#175 Removed from "Mentions" type unused fields and renamed files.
