+++ lib/types/configTypes.ts
